implementation authorisation using first GATE (comment after) and using in second time POLICIES (the one which is use)

+++ app/Http/Controllers/QuestionsController.php
@@ -8,6 +8,9 @@ use App\Http\Requests\AskQuestionRequest;
 
 class QuestionsController extends Controller
 {
+    public function __construct(){
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -67,7 +70,13 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        $this->authorize("update", $question);
+        //if (\Gate::denies('update-question', $question)){
+          //  abort(403, "Access denied");
+        //}
         return view("questions.edit", compact('question'));
+
+
     }
 
     /**
@@ -79,6 +88,10 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        $this->authorize("update", $question);
+        //if (\Gate::denies('update-question', $question)){
+          //  abort(403, "Access denied");
+        //}
         $question->update($request->only('title', 'body'));
         return redirect('/questions')->with('success', "Your question has been update");
     }
@@ -91,6 +104,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        $this->authorize("delete", $question);
+        //if (\Gate::denies('delete-question', $question)){
+        //    abort(403, "Access denied");
+        //}
         $question->delete();
 
         return redirect('/questions')->with('success', "Your question has been deleted.");
